refactor: organize themed components into base directory
Moved ThemedText and ThemedView components to components/base/ for better
component organization and updated all import paths across the codebase.
