+++ src/components/HelloWave.tsx
@@ -8,7 +8,7 @@ import Animated, {
   withTiming,
 } from 'react-native-reanimated';
 
-import { ThemedText } from '@/components/ThemedText';
+import { ThemedText } from '@/components/base/ThemedText';
 
 export function HelloWave() {
   const rotationAnimation = useSharedValue(0);
